# docs/swagger.yaml
basePath: /cmd
definitions:
  promo.Grade:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - trainee
    - junior
    - middle
    - senior
  promo.User:
    properties:
      id:
        type: integer
      name:
        type: string
      position:
        $ref: '#/definitions/promo.Grade'
      project:
        type: string
      surname:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API server for Andersen Promo Application
  title: Andersen Promo API
  version: "1.0"
paths:
  /create:
    post:
      consumes:
      - application/json
      description: set new user
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/promo.User'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create new user
      tags:
      - users
  /delete/{id}:
    delete:
      description: remove user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete user
      tags:
      - users
  /get/{id}:
    get:
      description: get user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/promo.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get user
      tags:
      - users
  /getall:
    get:
      description: get users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/promo.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List users
      tags:
      - users
  /healthcheck:
    get:
      description: get availability status
      responses:
        "200":
          description: OK
      summary: Checking availability
      tags:
      - users
  /update/{id}:
    patch:
      consumes:
      - application/json
      description: change user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/promo.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update user
      tags:
      - users
swagger: "2.0"